Refactor command modules into their own package.
