+++ iscc_cli/commands/batch.py
